BS/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
paddlepaddle==2.6.2
paddleocr
simple-lama-inpainting
python-pptx
Pillow
torch
torchvision
numpy==1.26.4
